Extract ProxyStartCommand from ProxyCommand.Start inner class
Move the Vert.x-dependent proxy start logic into a top-level
ProxyStartCommand class, isolating all Vert.x/Netty/MitmProxy imports
from the rest of ProxyCommand. This is preparation for the multi-module
split where the proxy binary and CLI/TUI binary have different
dependency sets.
